# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rocky Mountain Race Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /riders:
    get:
      operationId: listRiders
      tags: [riders]
      summary: List riders
      parameters:
        - name: clubId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rider"
    post:
      operationId: createRider
      tags: [riders]
      summary: Create a rider record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderInput"
      responses:
        "201":
          description: Rider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
  /riders/{riderId}:
    get:
      operationId: getRider
      tags: [riders]
      summary: Get rider details with history
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rider details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderDetail"
    patch:
      operationId: updateRider
      tags: [riders]
      summary: Update rider
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderUpdate"
      responses:
        "200":
          description: Rider updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"

  # REGISTRATIONS
...
  /rfid:
    get:
      operationId: listRfidTags
      tags: [rfid]
      summary: List RFID transponder assignments
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: RFID assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RfidAssignment"
    post:
      operationId: assignRfid
      tags: [rfid]
      summary: Assign RFID transponder to a rider for an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfidAssignmentInput"
      responses:
        "201":
          description: RFID assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfidAssignment"

  # MOTOS / HEATS
...
  /events/{eventId}/results:
    get:
      operationId: listResults
      tags: [results]
      summary: Get race results for an event (public)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RaceResult"
    post:
      operationId: submitResults
      tags: [results]
      summary: Submit race results for a moto
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultsInput"
      responses:
        "201":
          description: Results submitted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RaceResult"
  /events/{eventId}/results/publish:
    post:
      operationId: publishResults
      tags: [results]
      summary: Publish or unpublish event results
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishInput"
      responses:
        "200":
          description: Publication status updated

  # SERIES POINTS
  /series:
    get:
      operationId: listSeries
      tags: [series]
      summary: List all series
      responses:
        "200":
          description: Series list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"
    post:
      operationId: createSeries
      tags: [series]
      summary: Create a series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeriesInput"
      responses:
        "201":
          description: Series created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
  /series/{seriesId}/leaderboard:
    get:
      operationId: getSeriesLeaderboard
      tags: [series]
      summary: Get series leaderboard (public)
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Series leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeriesStanding"
  /series/{seriesId}/recalculate:
    post:
      operationId: recalculateSeriesPoints
      tags: [series]
      summary: Recalculate series points from results
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Points recalculated

  # DASHBOARD / WOW ENDPOINTS
...
components:
  schemas:
...
    Rider:
      type: object
      required: [id, firstName, lastName]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        bibNumber:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        rfidNumber:
          type: ["string", "null"]
        createdAt:
          type: string
    RiderDetail:
      type: object
      required: [id, firstName, lastName]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        bibNumber:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        rfidNumber:
          type: ["string", "null"]
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/RaceResult"
        totalEvents:
          type: integer
        createdAt:
          type: string
    RiderInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        bibNumber:
          type: string
        dateOfBirth:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        rfidNumber:
          type: string
    RiderUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        bibNumber:
          type: string
        dateOfBirth:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        rfidNumber:
          type: string
...
    RfidAssignment:
      type: object
      required: [id, riderId, rfidNumber]
      properties:
        id:
          type: integer
        riderId:
          type: integer
        riderName:
          type: string
        rfidNumber:
          type: string
        eventId:
          type: ["integer", "null"]
        assignedAt:
          type: string
    RfidAssignmentInput:
      type: object
      required: [riderId, rfidNumber]
      properties:
        riderId:
          type: integer
        rfidNumber:
          type: string
        eventId:
          type: integer
...
    RaceResult:
      type: object
      required: [id, eventId, motoId, riderId, riderName, position]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        motoId:
          type: integer
        motoName:
          type: string
        riderId:
          type: integer
        riderName:
          type: string
        raceClass:
          type: string
        position:
          type: integer
        totalTime:
          type: ["string", "null"]
        lapTimes:
          type: array
          items:
            type: string
        points:
          type: ["integer", "null"]
        dnf:
          type: boolean
        dns:
          type: boolean
        bibNumber:
          type: ["string", "null"]
    ResultsInput:
      type: object
      required: [motoId, results]
      properties:
        motoId:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/RiderResultInput"
    RiderResultInput:
      type: object
      required: [riderId, position]
      properties:
        riderId:
          type: integer
        position:
          type: integer
        totalTime:
          type: string
        lapTimes:
          type: array
          items:
            type: string
        dnf:
          type: boolean
        dns:
          type: boolean
    PublishInput:
      type: object
      required: [published]
      properties:
        published:
          type: boolean
    Series:
      type: object
      required: [id, name, clubId, season]
      properties:
        id:
          type: integer
        name:
          type: string
        clubId:
          type: integer
        season:
          type: integer
        classes:
          type: array
          items:
            type: string
        pointsSystem:
          type: string
        eventIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
    SeriesInput:
      type: object
      required: [name, clubId, season]
      properties:
        name:
          type: string
        clubId:
          type: integer
        season:
          type: integer
        classes:
          type: array
          items:
            type: string
        pointsSystem:
          type: string
        eventIds:
          type: array
          items:
            type: integer
    SeriesStanding:
      type: object
      required: [position, riderId, riderName, raceClass, totalPoints]
      properties:
        position:
          type: integer
        riderId:
          type: integer
        riderName:
          type: string
        raceClass:
          type: string
        totalPoints:
          type: integer
        eventsEntered:
          type: integer
        eventResults:
          type: array
          items:
            type: integer
...