# lib/api-zod/src/generated/types/riderDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rocky Mountain Race Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RaceResult } from './raceResult';

export interface RiderDetail {
  id: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bibNumber?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
  /** @nullable */
  emergencyPhone?: string | null;
  /** @nullable */
  rfidNumber?: string | null;
  recentResults?: RaceResult[];
  totalEvents?: number;
  createdAt?: string;
}